Final-ish(hehe) version of crow
Crow is working as i want it to work. Did not use OnPathComplete because i think it is triggered not when the whole path is complete but when the object reaches any waypoint leading to the path. Because of that usage of OnPathComplete in this specific scenario is not optimal